requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.24.3

# LinkedIn scraping
linkedin-api==2.0.0
python-linkedin-v2==0.9.4

# NLP and ML
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1
openai==1.3.7
sentence-transformers==2.2.2

# Data processing
python-dotenv==1.0.0
pydantic==2.5.0
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23

# Utilities
tqdm==4.66.1
loguru==0.7.2
click==8.1.7
rich==13.7.0 